Use JSON format for panel configuration
So now we use jansson library to handle panel configuration, instead of reinventing the wheel. That approach has several advantages. First, the code becomes much more readable. Second, the configuration is stored internally as JSON nodes, and the application just ignores items it does not understand. That allows us to get better compatability between future versions of the application: unknown options are not lost on saving the configuration to the file, so different versions can be used alternately. And third, the configuration file can be easily handled and changed by user scripts and third-party applications.

This commit applies new API to the following plugins: taskbar, dclock, launchbutton, pager, separator, space, tray, wincmd.